# AsynCUDA13.Client/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "AsynCUDA13.API v1",
    "version": "v1"
  },
  "paths": {
...
    "/kernels": {
      "get": {
        "tags": [
          "RuntimeKernel"
        ],
        "parameters": [
          {
            "name": "filterCompiled",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuntimeKernelInfo"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuntimeKernelInfo"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuntimeKernelInfo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/compile": {
      "post": {
        "tags": [
          "RuntimeKernel"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeCompileRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeCompileRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeCompileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeCompileResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeCompileResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeCompileResponse"
                }
              }
            }
          }
        }
      }
    },
    "/execute-generic": {
      "post": {
        "tags": [
          "RuntimeKernel"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeExecuteRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeExecuteRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeExecuteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeExecuteResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeExecuteResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeExecuteResponse"
                }
              }
            }
          }
        }
      }
    },
    "/execute-linear": {
      "post": {
        "tags": [
          "RuntimeKernel"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeExecuteRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeExecuteRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeExecuteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeExecuteResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeExecuteResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeExecuteResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IntPtr": {
        "type": "object",
        "additionalProperties": false
      },
...
      "RuntimeCompileRequest": {
        "required": [
          "kernelName",
          "kernelSource"
        ],
        "type": "object",
        "properties": {
          "kernelName": {
            "type": "string",
            "nullable": true
          },
          "kernelSource": {
            "type": "string",
            "nullable": true
          },
          "silent": {
            "type": "boolean"
          },
          "asyncCall": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "RuntimeCompileResponse": {
        "type": "object",
        "properties": {
          "kernelInfo": {
            "$ref": "#/components/schemas/RuntimeKernelInfo"
          },
          "buildLog": {
            "type": "string",
            "nullable": true
          },
          "elapsedMs": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "RuntimeExecuteRequest": {
        "required": [
          "kernelInfo"
        ],
        "type": "object",
        "properties": {
          "kernelInfo": {
            "$ref": "#/components/schemas/RuntimeKernelInfo"
          },
          "argumentValues": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "asyncCall": {
            "type": "boolean"
          },
          "unloadAfterExecution": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "RuntimeExecuteResponse": {
        "required": [
          "kernelInfo"
        ],
        "type": "object",
        "properties": {
          "resultPointer": {
            "$ref": "#/components/schemas/IntPtr"
          },
          "success": {
            "type": "boolean"
          },
          "kernelInfo": {
            "$ref": "#/components/schemas/RuntimeKernelInfo"
          },
          "elapsedMs": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "RuntimeKernelInfo": {
        "type": "object",
        "properties": {
          "functionName": {
            "type": "string",
            "nullable": true
          },
          "sourcePath": {
            "type": "string",
            "nullable": true
          },
          "ptxPath": {
            "type": "string",
            "nullable": true
          },
          "kernelCode": {
            "type": "string",
            "nullable": true
          },
          "argumentNames": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "argumentTypes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}