# AsynCUDA13.Shared/Api/Requests/RuntimeExecuteRequest.cs
using AsynCUDA13.Shared.RuntimeDtos;
using System;
using System.Collections.Generic;
using System.Text;

namespace AsynCUDA13.Shared.Api.Requests
{
    public class RuntimeExecuteRequest
    {
        public required RuntimeKernelInfo KernelInfo { get; set; }

        public IEnumerable<string> ArgumentValues { get; set; } = [];   // Arg values ToString() and in correct order

        public bool AsyncCall { get; set; } = true; // Async execution

        public bool UnloadAfterExecution { get; set; } = false; // Unload kernel after execution
    }
}